оч много всего

+++ Kanban/Domain/Chat.cs
@@ -2,12 +2,12 @@
 
 namespace Domain
 {
-    public class Chat : IEntity<int>
+    public class Chat : IEntity<long>
     {
-        public int Id { get; }
+        public long Id { get; }
         public App App { get; }
         public string ProjectId { get; }
 
-        public Chat(int id, App app, string projectId) => (Id, App, ProjectId) = (id, app, projectId);
+        public Chat(long id, App app, string projectId) => (Id, App, ProjectId) = (id, app, projectId);
     }
 }
